+++ packages/apollo-server-core/src/runHttpQuery.ts
@@ -186,8 +186,6 @@ export async function runHttpQuery(
     debug: options.debug,
 
     plugins: options.plugins || [],
-
-    reporting: options.reporting,
   };
 
   return processHTTPRequest(config, request);
@@ -265,7 +263,8 @@ export async function processHTTPRequest<TContext>(
       cache: options.cache,
       debug: options.debug,
       metrics: {
-        captureTraces: !!options.reporting,
+        // Defaults to false. Gets set to true by the reporting agent
+        captureTraces: false,
       },
     };
   }